+++ memory-viz/src/tests/draw.spec.tsx
@@ -20,7 +20,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -30,7 +30,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "bool", id: 32, value: true }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -40,7 +40,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "int", id: 32, value: 7 }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -50,7 +50,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "float", id: 32, value: 7.0 }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -62,7 +62,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -74,7 +74,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -84,7 +84,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "set", id: 32, value: [] }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -101,7 +101,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -113,7 +113,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -123,7 +123,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "list", id: 32, value: [] }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -140,7 +140,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -152,7 +152,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -162,7 +162,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "tuple", id: 32, value: [] }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -178,7 +178,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -188,7 +188,7 @@ describe("draw function", () => {
         const objects: Array<Object> = [{ type: "dict", id: 32, value: {} }];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -200,7 +200,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -212,7 +212,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -236,7 +236,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -257,7 +257,7 @@ describe("draw function", () => {
             },
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, false, {
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -274,7 +274,7 @@ describe("draw function", () => {
             },
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, false, {
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -287,7 +287,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -313,7 +313,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -347,7 +347,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -386,7 +386,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -410,7 +410,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -436,7 +436,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -462,7 +462,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -488,7 +488,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -514,7 +514,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -540,7 +540,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -566,7 +566,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -592,7 +592,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -618,7 +618,7 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
@@ -644,11 +644,12 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();
     });
+
     it("renders combinations of style presets", () => {
         const objects: Array<Object> = [
             {
@@ -669,7 +670,81 @@ describe("draw function", () => {
         ];
         const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
             width: 1300,
-            seed: 12345,
+            roughjs_config: { options: { seed: 12345 } },
+        });
+        const svg: String = m.serializeSVG();
+        expect(svg).toMatchSnapshot();
+    });
+
+    it("renders diagrams with provided roughjs_config 'fill' option", () => {
+        const objects: Array<Object> = [
+            {
+                type: "str",
+                id: 42,
+                value: "hello",
+            },
+        ];
+        const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
+            width: 1300,
+            roughjs_config: { options: { fill: "red", seed: 12345 } },
+        });
+        const svg: String = m.serializeSVG();
+        expect(svg).toMatchSnapshot();
+    });
+
+    it("renders diagrams with provided roughjs_config 'fill' and 'fillStyle' options", () => {
+        const objects: Array<Object> = [
+            {
+                type: "str",
+                id: 42,
+                value: "hello",
+            },
+        ];
+        const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
+            width: 1300,
+            roughjs_config: {
+                options: { fill: "green", fillStyle: "dashed", seed: 12345 },
+            },
+        });
+        const svg: String = m.serializeSVG();
+        expect(svg).toMatchSnapshot();
+    });
+
+    it("renders diagrams with provided roughjs_config 'roughness' option", () => {
+        const objects: Array<Object> = [
+            {
+                type: "str",
+                id: 42,
+                value: "hello",
+            },
+        ];
+        const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
+            width: 1300,
+            roughjs_config: { options: { roughness: 4, seed: 12345 } },
+        });
+        const svg: String = m.serializeSVG();
+        expect(svg).toMatchSnapshot();
+    });
+
+    it("renders diagrams with provided mix of roughjs_config options", () => {
+        const objects: Array<Object> = [
+            {
+                type: "str",
+                id: 42,
+                value: "hello",
+            },
+        ];
+        const m: InstanceType<typeof MemoryModel> = draw(objects, true, {
+            width: 1300,
+            roughjs_config: {
+                options: {
+                    roughness: 4,
+                    bowing: 5,
+                    fill: "blue",
+                    fillWeight: 5,
+                    seed: 12345,
+                },
+            },
         });
         const svg: String = m.serializeSVG();
         expect(svg).toMatchSnapshot();